Explicitly specify integer dtype
This should fix numpy 2.x compatibility on Windows

+++ src/libertem_blobfinder/udf/integration.py
@@ -32,13 +32,13 @@ class IntegrationUDF(UDF):
         >>> peaks_shape = nav_shape + extra_shape
 
         >>> # Generate some random positions as an example
-        >>> peaks = np.random.randint(low=0, high=np.min(sig_shape), size=peaks_shape)
+        >>> peaks = np.random.randint(low=0, high=np.min(sig_shape), size=peaks_shape, dtype=np.int64)
 
         >>> # Create an AuxBufferWrapper for the peaks
         >>> centers = IntegrationUDF.aux_data(
         ...     data=peaks,
         ...     kind='nav',
-        ...     dtype=int,
+        ...     dtype=np.int64,
         ...     extra_shape=extra_shape
         ... )
